# faster `hypot(::IEEEFloat...)` (re-do)