Redoing #48130 but without the `@fastmath` that was causing build problems. Performance is not quite as good as that PR but remains comparable for small numbers of inputs. The performance loss will be recovered if #48129 is resolved. In any case, it's a big improvement over master on an x86-64 architecture.

```julia-repl
using BenchmarkTools

t3 = ntuple(_->randn(), 3)
t4 = ntuple(_->randn(), 4)
t8 = ntuple(_->randn(), 8)
t16 = ntuple(_->randn(), 16)

@btime hypot($t3...)
@btime hypot($t4...)
@btime hypot($t8...)
@btime hypot($t16...)

# master
  9.810 ns (0 allocations: 0 bytes)
  10.110 ns (0 allocations: 0 bytes)
  19.759 ns (0 allocations: 0 bytes)
  50.203 ns (0 allocations: 0 bytes)

# this PR
  5.500 ns (0 allocations: 0 bytes)
  7.608 ns (0 allocations: 0 bytes)
  11.411 ns (0 allocations: 0 bytes)
  19.960 ns (0 allocations: 0 bytes)
```

The biggest difference is a replacing `maximum(abs,x)` with a specialized version for `IEEEFloats`. This performance difference is significant on x86-64 but may not exist on aarch64 since it has a native `max` instruction. This also uses fast scaling values (rather than division) to achieve additional speedup.