@mikmoore 2023-02-11
At @oscardssmith's suggestion, I've assembled a change. Using power-of-two normalization allows for this algorithm to avoid division operations. The speedup is significant for 3 arguments but is small after that. The `code_native` did not offer an explanation - I can only assume that more inputs allowed it to bury the latency of the divisions.

```julia-repl
## see original post for benchmark

# previous timings
  8.008 ns (0 allocations: 0 bytes)
  7.508 ns (0 allocations: 0 bytes)
  12.613 ns (0 allocations: 0 bytes)
  21.565 ns (0 allocations: 0 bytes)

# new timings
  5.500 ns (0 allocations: 0 bytes)
  7.608 ns (0 allocations: 0 bytes)
  11.411 ns (0 allocations: 0 bytes)
  19.960 ns (0 allocations: 0 bytes)
```

This change includes a new internal function, currently called `Base.Math.scaleinv`. I am open to suggestions for naming, placement, implementation, etc. I am making this function available because it came up in #48563 as well as in some upcoming work I'm doing on other operations.