@mikmoore 2023-02-19
New timings with fastmath reinstated:
```julia-repl
# master
  9.810 ns (0 allocations: 0 bytes)
  10.110 ns (0 allocations: 0 bytes)
  19.759 ns (0 allocations: 0 bytes)
  50.203 ns (0 allocations: 0 bytes)

# pre-fastmath
  5.500 ns (0 allocations: 0 bytes)
  7.608 ns (0 allocations: 0 bytes)
  11.411 ns (0 allocations: 0 bytes)
  19.960 ns (0 allocations: 0 bytes)

# with fastmath
  5.400 ns (0 allocations: 0 bytes)
  7.708 ns (0 allocations: 0 bytes)
  9.510 ns (0 allocations: 0 bytes)
  13.928 ns (0 allocations: 0 bytes)
```